todolist-support/challenge/requirements.txt
argon2-cffi==21.3.0
cryptography==37.0.2
Flask==2.1.2
flask-redis==0.4.0
Flask-Session==0.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
passlib==1.7.4
PyJWT==2.4.0
redis==4.3.1
scrypt==0.8.20
Werkzeug==2.1.2
requests==2.27.1
furl==2.1.3
